shell-test.sh library added TEST_FAILURES counter and ERROR_STOP value to allow test plans to carry on if out/base file comparison fails.

+++ bin/tests/filter-css-colors/tests.sh
@@ -13,6 +13,8 @@ SKIP=0
 source ../shell-test.sh
 [ -d out ] || mkdir out
 rm out/* > /dev/null 2>&1 || echo OK output dir ready
+# Do not terminate test plan if out/base comparison fails.
+ERROR_STOP=0
 
 echo TEST --version option
 TEST=version-option
@@ -222,5 +224,7 @@ else
 fi
 
 # clean up output directory if no failures
-rm out/* && rmdir out
-echo OK All tests complete `pwd`/out cleaned up
+if [ $TEST_FAILURES == 0 ]; then
+	rm out/* && rmdir out
+	echo OK All tests complete `pwd`/out cleaned up
+fi
